foundation-project/backend/requirements.txt
Flask
Flask-CORS
boto3
python-dotenv
Werkzeug

